backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "firebase-admin>=7.4.0",
    "huggingface-hub>=1.11.0",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "transformers>=5.5.4",
    "ultralytics>=8.4.39",
    "uvicorn>=0.44.0",
]
